Working on moving to one container for unit storage

+++ Source/McRave/GridManager.cpp
@@ -402,14 +402,14 @@ namespace McRave
 		int speed = int(max(unit.getSpeed(), 1.0));
 
 		int pixelSize = max(unit.getType().width(), unit.getType().height());
-		int walkSize = ceil(pixelSize / 8);
+		int walkSize = int(ceil(pixelSize / 8));
 
 		int grdReach = int(max(unit.getGroundRange(), 32.0) + (speed * 24.0) + (pixelSize / 2)) + 1;
 		int airReach = int(max(unit.getAirRange(), 32.0) + (speed * 24.0) + (pixelSize / 2)) + 1;
 
 		if (unit.getType().isWorker()) {
-			grdReach = grdReach / 1.5;
-			airReach = airReach / 1.5;
+			grdReach = int(grdReach / 1.5);
+			airReach = int(airReach / 1.5);
 		}
 
 		int radius = 1 + (max(grdReach, airReach))/8;
